+++ libs/dot-components/src/lib/components/avatar/Avatar.tsx
@@ -2,7 +2,7 @@ import React, { MouseEvent } from 'react';
 import { Typography } from '@material-ui/core';
 
 import { CommonProps } from '../CommonProps';
-import { useStylesWithRootClass } from '../makeStylesWithRootClass';
+import { useStylesWithRootClass } from '../useStylesWithRootClass';
 import { DotIcon } from '../icon/Icon';
 import { rootClassName, StyledAvatar } from './Avatar.styles';