+++ libs/dot-components/src/lib/components/drawer/Drawer.styles.ts
@@ -1,11 +1,11 @@
-import { Drawer, Theme, fade } from '@material-ui/core';
+import { Drawer, fade } from '@material-ui/core';
 import styled, { css } from 'styled-components';
 import { DrawerProps } from './Drawer';
 
 export const rootClassName = 'dot-drawer';
 
 export const StyledDrawer = styled(Drawer)<DrawerProps>`
-  ${({ theme }: { theme: Theme }) => css`
+  ${({ theme }) => css`
     &.dot-drawer {
       .MuiBackdrop-root {
         background-color: ${fade(theme.palette.grey[900], 0.7)};