+++ libs/dot-components/src/lib/components/input-form-fields/InputText.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { renderWithTheme as render } from '../../testing-utils/RenderWithTheme';
 import { screen } from '@testing-library/dom';
 import userEvent from '@testing-library/user-event';
-import { DotInputText } from './InputFormFields';
+import { DotInputText } from './InputText';
 
 describe('DotInputText', () => {
   it('renders successfully', () => {