+++ libs/dot-components/src/lib/components/makeStylesWithRootClass.ts
@@ -1,6 +1,0 @@
-export function useStylesWithRootClass(name: string, className?: string) {
-  // combine the root classes into one string
-  const rootClasses = [name, ...(className ? [className] : [])].join(' ');
-
-  return rootClasses;
-}