+++ CHANGE_LOG.md
@@ -12,6 +12,22 @@
 - **EmptyState**: changed `Image` prop to `image`
 - **Navigation**: added button as component to trigger subnav
 - **Navigation**: added support for keyboard navigation
+- **InputText**: added props for `startIcon`, `endIcon`, `warning`, `id`, `helperText`
+- **InputText**: added custom styles for `warning` state
+- **InputText**: fixed inconnsistent styles per UX review
+- **InputText**: added better demo controls in storybook for `startIcon`, `endIcon`
+- **InputText**: moved styled component definitions to `InputFormFields.styles.ts`
+- **InputText**: moved interface and type definitions to `InputFormFields.propTypes.ts`
+- **InputText**: removed `variant` prop
+- **InputSelect**: added props for `startIcon`, `endIcon`, `warning`, `id`, `helperText`
+- **InputSelect**: added custom styles for `warning` state
+- **InputSelect**: fixed inconnsistent styles per UX review
+- **InputSelect**: added better demo controls in storybook for `startIcon`, `endIcon`
+- **InputSelect**: moved styled component definitions to `InputFormFields.styles.ts`
+- **InputSelect**: moved interface and type definitions to `InputFormFields.propTypes.ts`
+- **InputSelect**: removed `variant` prop
+- **AutoComplete**: removed `inputVariant` prop
+- **InlineEdit**: removed `inputVariant` prop
 
 ## 0.0.0-alpha.3